+++ source/Application/views/admin/de/lang.php
@@ -1937,11 +1937,6 @@ $aLang = [
     'mxo3captcha' => 'CAPTCHA',
     'tbclo3captcha_config' => 'Einstellungen',
     'O3_CAPTCHA_PROVIDER_NONE' => 'Keiner (deaktiviert)',
-    'O3_CAPTCHA_PROVIDER_GOOGLE_V2' => 'Google reCAPTCHA v2 (checkbox)',
-    'O3_CAPTCHA_PROVIDER_GOOGLE_V3' => 'Google reCAPTCHA v3 (invisible/score)',
-    'O3_CAPTCHA_SITE_KEY' => 'Site Key',
-    'O3_CAPTCHA_SECRET_KEY' => 'Secret Key',
-    'O3_CAPTCHA_SCORE_THRESHOLD' => 'Mindest-Score (0.0–1.0)',
     'O3_CAPTCHA_REQUIRE_CONSENT' => 'Einwilligung des Besuchers vor dem Laden erforderlich',
     'O3_CAPTCHA_FORM_contact' => 'Kontaktformular',
     'O3_CAPTCHA_FORM_newsletter' => 'Newsletter-Formular',
