+++ source/Application/views/admin/en/lang.php
@@ -1936,11 +1936,6 @@ $aLang = [
     'mxo3captcha' => 'CAPTCHA',
     'tbclo3captcha_config' => 'Settings',
     'O3_CAPTCHA_PROVIDER_NONE' => 'None (disabled)',
-    'O3_CAPTCHA_PROVIDER_GOOGLE_V2' => 'Google reCAPTCHA v2 (checkbox)',
-    'O3_CAPTCHA_PROVIDER_GOOGLE_V3' => 'Google reCAPTCHA v3 (invisible/score)',
-    'O3_CAPTCHA_SITE_KEY' => 'Site key',
-    'O3_CAPTCHA_SECRET_KEY' => 'Secret key',
-    'O3_CAPTCHA_SCORE_THRESHOLD' => 'Minimum score to pass (0.0–1.0)',
     'O3_CAPTCHA_REQUIRE_CONSENT' => 'Require visitor consent before loading',
     'O3_CAPTCHA_FORM_contact' => 'Contact form',
     'O3_CAPTCHA_FORM_newsletter' => 'Newsletter form',
